backend/pyproject.toml
[project]
name = "investai-backend"
version = "0.1.0"
description = "InvestAI FastAPI backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.135.1",
    "uvicorn[standard]==0.41.0",
    "sqlalchemy[asyncio]==2.0.48",
    "alembic==1.18.4",
    "aiosqlite==0.22.1",
    "asyncpg==0.31.0",
    "pydantic-settings==2.13.1",
    "bcrypt>=4.0.1",
    "pyjwt==2.10.1",
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "httpx==0.28.1",
]
schwab = [
    "schwab-py==1.5.1",
    "python-dotenv==1.2.2",
    "pandas==3.0.1",
    "duckdb==1.5.0",
]

[build-system]
requires = ["hatchling==1.29.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
